+++ tests/unit/test_align_celeba_format.py
@@ -167,4 +167,4 @@ class AlignCelebaImporterTest(TestCase):
     @mark_requirement(Requirements.DATUM_475)
     def test_can_detect_align_dataset(self):
         detected_formats = Environment().detect_dataset(DUMMY_ALIGN_DATASET_DIR)
-        self.assertIn(AlignCelebaImporter.NAME, detected_formats)
+        self.assertEqual([AlignCelebaImporter.NAME], detected_formats)